demo/requirements.txt
eggai==0.1.46
python-dotenv==1.0.1
fastapi==0.115.6
dspy==2.6.24
psutil==7.0.0
websockets==14.1
guardrails-ai==0.6.2
pyvespa==0.56.0
tenacity==9.0.0
httpx==0.28.1
requests==2.32.2
uvicorn==0.34.0
openlit==1.33.21
opentelemetry-sdk==1.33.1
python-json-logger==2.0.7
colorlog==6.8.2
pydantic-settings==2.4.0
transformers[torch]>=4.51.3
peft>=0.14.0
datasets>=3.3.2
accelerate>=1.4.0
bitsandbytes>=0.42.0
trl>=0.15.2
numpy==2.2.4
scikit-learn==1.6.1
aioboto3==13.3.0
mlflow==2.21.3
matplotlib==3.10.1
seaborn==0.13.2
pandas==2.2.3
sentence-transformers==4.0.2
pytest-html==4.1.1
jupyter==1.1.1
aiofiles>=24.1.0
asyncio==3.4.3
litellm==1.70.2
pydantic
tqdm
jinja2>=3.1.3
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
temporalio>=1.8.0
ruff==0.11.11
pytest-cov==6.1.1
tabulate==0.9.0
docling==2.37.0